move edge store out of internal pkg to enable testing
