+++ central/componentcveedge/datastore/store/postgres/store.go
@@ -45,7 +45,6 @@ type Store interface {
 	Count(ctx context.Context) (int, error)
 	Exists(ctx context.Context, id string) (bool, error)
 	Get(ctx context.Context, id string) (*storage.ComponentCVEEdge, bool, error)
-	GetByQuery(ctx context.Context, query *v1.Query) ([]*storage.ComponentCVEEdge, error)
 	GetIDs(ctx context.Context) ([]string, error)
 	GetMany(ctx context.Context, ids []string) ([]*storage.ComponentCVEEdge, []int, error)
 
@@ -189,35 +188,6 @@ func (s *storeImpl) GetMany(ctx context.Context, ids []string) ([]*storage.Compo
 		}
 	}
 	return elems, missingIndices, nil
-}
-
-// GetByQuery returns the objects matching the query
-func (s *storeImpl) GetByQuery(ctx context.Context, query *v1.Query) ([]*storage.ComponentCVEEdge, error) {
-	defer metrics.SetPostgresOperationDurationTime(time.Now(), ops.GetByQuery, "ComponentCVEEdge")
-
-	var sacQueryFilter *v1.Query
-
-	q := search.ConjunctionQuery(
-		sacQueryFilter,
-		query,
-	)
-
-	rows, err := postgres.RunGetManyQueryForSchema(ctx, schema, q, s.db)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, nil
-		}
-		return nil, err
-	}
-	var results []*storage.ComponentCVEEdge
-	for _, data := range rows {
-		msg := &storage.ComponentCVEEdge{}
-		if err := proto.Unmarshal(data, msg); err != nil {
-			return nil, err
-		}
-		results = append(results, msg)
-	}
-	return results, nil
 }
 
 // Walk iterates over all of the objects in the store and applies the closure
