+++ central/graphql/resolvers/vulnerability_requests_test.go
@@ -248,7 +248,7 @@ func (s *VulnRequestResolverTestSuite) createImageDataStore(dacky *dackbox.DackB
 func (s *VulnRequestResolverTestSuite) createComponentCVEDataStore(dacky *dackbox.DackBox) {
 	componentCVEEdgeStore, err := componentCVEEdgeStore.New(dacky)
 	s.NoError(err)
-	s.componentCVEDataStore, err = componentCVEEdgeDS.New(dacky,
+	s.componentCVEDataStore = componentCVEEdgeDS.New(dacky,
 		componentCVEEdgeStore,
 		componentCVEEdgeIndex.New(s.bleveIndex),
 		componentCVEEdgeSearcher.New(componentCVEEdgeStore, dacky,
@@ -264,7 +264,6 @@ func (s *VulnRequestResolverTestSuite) createComponentCVEDataStore(dacky *dackbo
 			clusterIndex.New(s.bleveIndex),
 		),
 	)
-	s.NoError(err)
 }
 
 func (s *VulnRequestResolverTestSuite) createVulnRequestDataStore(pendingReqCache, activeReqCache vulnReqCache.VulnReqCache) {
